Finish Exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -44,7 +44,20 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({error: 'token invalid'})
+  }
+  const user = await User.findById(decodedToken.id)
+  
+  const blogToDelete = await Blog.findById(request.params.id)
+  if (decodedToken.id !== blogToDelete.user.toString()) {
+    return response.status(401).json({error: 'token invalid'})
+  } else {
+    await Blog.findByIdAndDelete(request.params.id)
+  }
+
   response.status(204).end()
 })
